requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==5.2.12
django-prometheus==2.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-writable-nested==0.7.2
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
flower==2.0.1
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.3.2
h11==0.16.0
humanize==4.15.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
locust==2.43.3
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.54.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
Werkzeug==3.1.6
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
